[FIX] 구독 후 리스트뷰 넘어갈 때 탭 스타일 적용
