+++ js/sections/mainView.js
@@ -1,4 +1,4 @@
-import { changeView } from "../utils/changeView.js";
+import { changeView, updateTabSelection } from "../utils/changeView.js";
 import { showGridView } from "../utils/makeGridView.js";
 import { showListView } from "../utils/makeListView.js";
 import {
@@ -9,13 +9,7 @@ import {
 } from "../constants/constants.js";
 import { store } from "../core/store.js";
 import { shuffleArray } from "../utils/shuffleIndex.js";
-import {
-  getView,
-  getPage,
-  getSubscribedPress,
-  getMode,
-  getCurrentPress,
-} from "../core/getter.js";
+import { getView, getPage, getMode, getCurrentPress } from "../core/getter.js";
 import { getData } from "../core/api.js";
 function MainView() {
   document.addEventListener("click", handleClick);
@@ -65,27 +59,6 @@ function drawPopup() {
   view_content.append(new_div_alert, new_div_snackbar);
 }
 
-function updateTabSelection(selectedTab) {
-  const allTab = document.getElementById("all");
-  const subscribeTab = document.getElementById("subscribe");
-
-  allTab.classList.remove("selected-bold16", "clicked");
-  allTab.classList.add("available-medium16");
-  subscribeTab.classList.remove("selected-bold16", "clicked");
-  subscribeTab.classList.add("available-medium16");
-  selectedTab.classList.remove("available-medium16");
-  selectedTab.classList.add("selected-bold16", "clicked");
-  store.setState({ page: FIRST_PAGE_NUM, tabMode: `${selectedTab.id}` });
-
-  if (getView() === "grid") {
-    showGridView();
-  } else {
-    showListView(
-      selectedTab.id === "all" ? CATEGORY[0] : getSubscribedPress()[0]
-    );
-  }
-}
-
 function handleClick(e) {
   const view_content = document.querySelector(".view-content");
   const target = e.target.id;
@@ -115,10 +88,8 @@ function handleClick(e) {
         : changePage(target);
       break;
     case "all":
-      updateTabSelection(document.getElementById("all"));
-      break;
     case "subscribe":
-      updateTabSelection(document.getElementById("subscribe"));
+      updateTabSelection(document.getElementById(`${target}`));
       break;
     default:
       break;
